requirements.txt
numpy==1.17.4
pandas==0.25.3
google_api_python_client==1.9.1
protobuf==3.12.2
